+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased][unreleased]
+### Added
+- added animated example script with Drawille chart on Paper decorator
+
 ### Changed
 - fixed IndexError in BoxRenderer & SharpRenderer